starting work to display shift time.  need to merge in dev changes to get bug fix.
