+++ Gateway/crds-angular/crds-angular.csproj
@@ -260,6 +260,7 @@
     <Compile Include="Models\Crossroads\FamilyMember.cs" />
     <Compile Include="Models\Crossroads\ProgramDTO.cs" />
     <Compile Include="Models\Crossroads\ServeOccurance.cs" />
+    <Compile Include="Models\Crossroads\Serve\Capacity.cs" />
     <Compile Include="Models\Crossroads\Serve\SaveRsvpDto.cs" />
     <Compile Include="Models\Crossroads\Serve\ServeRsvp.cs" />
     <Compile Include="Models\Crossroads\SignUpFamilyMembers.cs" />
